Create Apple and orange - Hackerrank (Easy)
Sam's house has an apple tree and an orange tree that yield an abundance of fruit. If we imagine Sam's house and garden as points on a numberline... the first line of inputs are s and t. Sam's hosue lies between these two points inclusively. The second line of inputs are a and b, the points of the apple tree (to the left of the house) and the orange tree (to the right of the house) respectively. The third line of inputs are m and n - the number of apples and oranges that have fallen from the tree. Fourth and fifth line of integers are lists of positions (relative to a and b) that the respective fruit has fallen (negative is left, positive is right). Your task is to first print how many apples will fall on the area covered by the house and then how many oranges will fall on this area.